refactor: split createGlobalState into its own function
